Added priority update event to sub-note component
