+++ components/pages/notes/note-card/index.vue
@@ -61,73 +61,13 @@
         />
       </label>
 
-      <div class="flex flex-col gap-2">
-        <div class="flex items-center justify-between px-6">
-          <span>Subnotes</span>
-
-          <NotesNoteCardSubNoteFilterPopover v-model="userSettings.filterOptions" />
-        </div>
-
-        <div class="flex flex-col">
-          <div v-if="loadingSubNotes && !subnotes.length">
-            <NotesNoteCardSubNoteSkeleton v-for="i in 5" :key="i" />
-          </div>
-
-          <div v-if="!loadingSubNotes && !openSubnotes.length" class="px-6">
-            <div class="text-sm opacity-50 italic">No open subnotes</div>
-          </div>
-
-          <VueDraggable
-            :disabled="userSettings.filterOptions.sortBy !== 'none'"
-            key="id"
-            :animation="150"
-            handle=".handle"
-            v-model="filteredAndSortedSubnotes"
-            @update="handleNotesOrder"
-          >
-            <NotesNoteCardSubNote
-              v-for="subnote in openSubnotes"
-              :key="subnote.id"
-              :note="subnote"
-              @deleted="handleRemoveSubnote(subnote)"
-              @add-after="handleAddNewAfter(subnote)"
-              :can-drag="userSettings.filterOptions.sortBy === 'none'"
-            />
-          </VueDraggable>
-        </div>
-
-        <div class="px-6">
-          <UButton label="Add note" icon="i-tabler-plus" :block="false" color="gray" variant="solid" @click="handleAddNote" />
-        </div>
-      </div>
-
-      <div class="flex flex-col gap-2">
-        <div class="flex items-center justify-between px-6">
-          <span>Completed subnotes</span>
-        </div>
-
-        <div class="flex flex-col">
-          <div v-if="loadingSubNotes && !subnotes.length">
-            <NotesNoteCardSubNoteSkeleton v-for="i in 5" :key="i" />
-          </div>
-
-          <NotesNoteCardSubNote
-            v-for="completedSubnote in completedSubnotes"
-            :key="completedSubnote.id"
-            :note="completedSubnote"
-            @deleted="handleRemoveSubnote(completedSubnote)"
-            @add-after="handleAddNewAfter(completedSubnote)"
-            :can-drag="false"
-          />
-        </div>
-      </div>
+      <NotesNoteCardSubNotes :note="note" />
     </div>
   </UCard>
 </template>
 
 <script setup lang="ts">
 import type { QueryData } from "@supabase/supabase-js";
-import { VueDraggable, type SortableEvent } from "vue-draggable-plus";
 import type { Tables } from "~/types/supabase";
 
 type BreakdcrumbLink = { id: number; label: string; to: string; icon?: string };
@@ -138,52 +78,7 @@ const props = defineProps<{
 
 const { projectId } = useParams();
 
-const { userSettings } = useUserSettings();
-
 const notesApi = useNotesApi();
-
-const subnotes = ref<Tables<"notes">[]>([]);
-const loadingSubNotes = ref(false);
-
-const filteredAndSortedSubnotes = ref<Tables<"notes">[]>([]);
-const openSubnotes = computed(() => filteredAndSortedSubnotes.value.filter((note) => !note.completed));
-const completedSubnotes = computed(() => filteredAndSortedSubnotes.value.filter((note) => note.completed));
-
-const filterAndSortSubnotes = () => {
-  const sortBy = userSettings.value.filterOptions.sortBy;
-  const sortOrder = userSettings.value.filterOptions.sortOrder;
-
-  if (sortBy === "none") return subnotes.value;
-
-  const handleSortByPriority = (a: Tables<"notes">, b: Tables<"notes">) => {
-    const aPriority = getPriotiryNumber(a.priority);
-    const bPriority = getPriotiryNumber(b.priority);
-
-    return sortOrder === "asc" ? aPriority - bPriority : bPriority - aPriority;
-  };
-
-  return subnotes.value.toSorted((a, b) => {
-    if (sortBy === "priority") {
-      return handleSortByPriority(a, b);
-    }
-
-    return 0;
-  });
-};
-
-const handleUpdateFilteredAndSortedSubnotes = () => {
-  filteredAndSortedSubnotes.value = filterAndSortSubnotes();
-};
-
-watch(
-  () => userSettings.value.filterOptions,
-  () => {
-    handleUpdateFilteredAndSortedSubnotes();
-  },
-  {
-    deep: true,
-  }
-);
 
 const title = ref(props.note.title ?? "");
 const description = ref(props.note.description ?? "");
@@ -246,59 +141,6 @@ const handleUpdateDescription = throttle(async (value: string) => {
   props.note.description = updatedNote.description;
 });
 
-const handleAddNote = async () => {
-  const newSubnote = await notesApi.createNoteFromNote(props.note.id, {
-    order: subnotes.value.length,
-  });
-  if (!newSubnote) return;
-  subnotes.value.push(newSubnote);
-  handleUpdateFilteredAndSortedSubnotes();
-};
-
-const loadSubNotes = async () => {
-  loadingSubNotes.value = true;
-  subnotes.value = await notesApi.getNotesOfParent(props.note.id);
-  handleUpdateFilteredAndSortedSubnotes();
-  loadingSubNotes.value = false;
-};
-
-const handleRemoveSubnote = (subnote: Tables<"notes">) => {
-  subnotes.value = subnotes.value.filter((note) => note.id !== subnote.id);
-  handleUpdateFilteredAndSortedSubnotes();
-};
-
-const handleAddNewAfter = async (subnote: Tables<"notes">) => {
-  const index = subnotes.value.findIndex((note) => note.id === subnote.id);
-  const newSubnote = await notesApi.createNoteFromNote(props.note.id);
-  if (!newSubnote) return;
-
-  subnotes.value.splice(index + 1, 0, newSubnote);
-  handleUpdateFilteredAndSortedSubnotes();
-
-  const notesToUpdate = subnotes.value.slice(index + 1);
-  const noteIds = notesToUpdate.map((note) => note.id);
-  const noteOrderIndexes = notesToUpdate.map((_note, i) => i + index);
-
-  notesApi.updateNotesOrder(noteIds, noteOrderIndexes);
-};
-
-const handleNotesOrder = async (event: SortableEvent) => {
-  const from = event.oldIndex;
-  const to = event.newIndex;
-
-  if (from === undefined || to === undefined) return;
-
-  const correctedFromIndex = Math.min(from, to);
-  const correctedToIndex = Math.max(from, to) + 1;
-
-  const notesToUpdate = filteredAndSortedSubnotes.value.slice(correctedFromIndex, correctedToIndex);
-
-  const noteIds = notesToUpdate.map((note) => note.id);
-  const noteOrders = notesToUpdate.map((_note, index) => correctedFromIndex + index);
-
-  notesApi.updateNotesOrder(noteIds, noteOrders);
-};
-
 const breadcrumbs = ref<BreakdcrumbLink[]>([]);
 const loadBreadcrumbs = async () => {
   const parents = await notesApi.loadAllNoteParents(props.note.id);
@@ -334,7 +176,6 @@ const loadBreadcrumbs = async () => {
   });
 };
 
-loadSubNotes();
 loadBreadcrumbs();
 
 const showDeleteModal = ref(false);
